requirements.txt
asdf>=1.2,<2.0
lz4>=0.10.1
numpy>=1.12,<2.0
scipy>=0.17,<1.0
clint>=0.5.0
google-cloud-storage>=1.0,<2.0
python-dateutil
